zazu/requirements.txt
black==22.6.0
pandas==1.4.3
mypy==0.971
pylint==2.14.5
pytest==7.1.2
python-dateutil==2.8.2
streamlit==1.12.0
